+++ packages/ui/index.d.ts
@@ -640,6 +640,7 @@ declare module "@elf-framework/ui" {
     gap: number;
   };
   export interface FlexProps {
+    class?: string;
     stack?: boolean;
     wrap?: boolean;
     style?: FlexStyle & CommonStyle;
@@ -665,6 +666,7 @@ declare module "@elf-framework/ui" {
 
   export type GridTemplate = number | string;
   export interface GridProps {
+    class: string;
     columns?: GridTemplate | GridTemplate[];
     rows?: GridTemplate | GridTemplate[];
     style?: GridStyle & CommonStyle;
@@ -997,4 +999,35 @@ declare module "@elf-framework/ui" {
   export class Divider extends UIElement {
     props: DividerProps;
   }
+
+  interface AlertProps {
+    type: "default" | "primary" | "success" | "info" | "warning" | "danger";
+    title: ContentType;
+    content: ContentType;
+    style: CommonStyle;
+    closable: boolean;
+    delay?: number;
+    onShow?: () => void;
+    onHide?: () => void;
+  }
+  export class Alert extends UIElement {
+    props: AlertProps & DomEventType;
+    hide(): void;
+  }
+
+  /** call Alert with potal */
+  export function alert(arg: {
+    content?: ContentType;
+    title?: ContentType;
+    delay?: number;
+    style?: CommonStyle;
+    closable?: boolean;
+    onShow?: () => void;
+    onHide?: () => void;
+    options?:
+      | {
+          container: HTMLElement | string;
+        }
+      | unknown;
+  }): Alert;
 }