+++ packages/design-tokens/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @elf-framework/design-tokens
 
+## 2.0.37
+
+### Patch Changes
+
+- Support component token (divider, alert, button)
+
 ## 2.0.35
 
 ### Patch Changes